refactor: Split up `core.hpp`
Move serialized (bus) message types to their own header in order to
break the `sink.hpp` -> `core.hpp` dependency.

Rename `dlog::core` to `dlog::log_fabric` in order to more aptly
describe its role.
