+++ libs/deeplog/src/dplx/dlog/core/serialized_messages.test.cpp
@@ -1,0 +1,8 @@
+
+// Copyright Henrik Steffen Gaßmann 2023
+//
+// Distributed under the Boost Software License, Version 1.0.
+//         (See accompanying file LICENSE or copy at
+//           https://www.boost.org/LICENSE_1_0.txt)
+
+#include "dplx/dlog/core/serialized_messages.hpp"
